testare

+++ app/lib/screens/screen_gestione/manage_activities.dart
@@ -3,6 +3,7 @@ import 'package:flutter_app2/screens/navdrawer_admin.dart';
 import '../../models/utente.dart';
 import '../../models/attivita.dart';
 import '../../repository/data_repository.dart';
+import 'package:flutter_email_sender/flutter_email_sender.dart';
 
 DataRepository repository = DataRepository();
 
@@ -24,7 +25,10 @@ class _ManageActivitiesState extends State<ManageActivities> {
         drawer: NavDrawerAdmin(),
       body: Padding(
           padding: const EdgeInsets.all(20),
-        child: ListView(
+          child: Container(
+          width: double.infinity,
+          height: 570,
+        child: Column(
             children: <Widget>[
               Container(
                 padding: const EdgeInsets.fromLTRB(10, 0, 10, 0),
@@ -49,11 +53,13 @@ class _ManageActivitiesState extends State<ManageActivities> {
                 margin: const EdgeInsets.fromLTRB(10, 30, 10, 0),
                 child: Text("Lista delle attività giornaliere", style: Theme.of(context).textTheme.headline6),
               ),
-              activityWidget(),
+              Expanded(
+                child:activityWidget(),
+              )
             ],
-        )),
+        ))),
       floatingActionButton: FloatingActionButton(
-        onPressed: () { Navigator.pushNamed(context, '/insert_activity').then((_) => setState(() {}));},
+        onPressed: () { Navigator.pushNamed(context, '/insert_activity', arguments: selectedDate).then((_) => setState(() {}));},
         child: Icon(Icons.add),
       ),
     );
@@ -85,8 +91,9 @@ class _ManageActivitiesState extends State<ManageActivities> {
           return Center(child: CircularProgressIndicator());
         }
         return ListView.builder(
-          shrinkWrap: true,
-          physics: ScrollPhysics(),
+          physics: ScrollPhysics(), ///
+          shrinkWrap: true, ///
+          scrollDirection: Axis.vertical,
           itemCount: snapshot.data.length,
           itemBuilder: (context, index) {
             Attivita attivita = snapshot.data.keys.elementAt(index);
@@ -102,20 +109,39 @@ class _ManageActivitiesState extends State<ManageActivities> {
                           //mainAxisAlignment: MainAxisAlignment.start,
                           children: [
                             Row(
-                                children: [Text("Ora inizio: "+ (attivita.orainizio?.hour.toString() ?? "")+":"+(attivita.orainizio?.minute.toString() ?? "")
-                              + "\nOra fine: " +  (attivita.orafine?.hour.toString() ?? "")+":"+(attivita.orafine?.minute.toString() ?? ""),
+                                children: [Text("Ora inizio: "+ (attivita.orainizio?.hour.toString().padLeft(2, '0') ?? "")+":"+(attivita.orainizio?.minute.toString().padLeft(2, '0') ?? "")
+                              + "\nOra fine: " +  (attivita.orafine?.hour.toString().padLeft(2, '0') ?? "")+":"+(attivita.orafine?.minute.toString().padLeft(2, '0') ?? ""),
                               )
                               ],
                             ),
                             Row(
                               children: [
                                 TextButton(child: Text('Modifica'), onPressed: () {
-                                    Navigator.pushNamed(context, '/modify_activity', arguments: [utente, attivita]).then((_) => setState(() {}));
+                                    Navigator.pushReplacementNamed(context, '/modify_activity', arguments: [utente, attivita]).then((_) => setState(() {}));
                                 }),
                                 TextButton(child: Text('Elimina'), onPressed: () {
-                                  _deleteActivity(attivita, utente);
-                                  setState(() {
-                                  });
+                                  showDialog(
+                                    context: context,
+                                    builder: (context) =>
+                                        AlertDialog(
+                                          title: Text(
+                                              "Sei sicuro di voler eliminare l'attività?"),
+                                          actions: <Widget>[
+                                            new ElevatedButton(
+                                              onPressed: () => Navigator.pop(context), // Closes the dialog
+                                              child: new Text('No'),
+                                            ),
+                                            new ElevatedButton(
+                                              onPressed: () {
+                                                Navigator.pop(context);
+                                                _deleteActivity(attivita, utente);
+                                                setState(() {});
+                                              },
+                                              child: new Text('Sì'),
+                                            ),
+                                          ],
+                                        ),
+                                  );
                                 })
                               ],
                             )
@@ -141,6 +167,15 @@ class _ManageActivitiesState extends State<ManageActivities> {
   _deleteActivity(Attivita a, Utente u){
     u.deleteActivity(a);
     repository.updateUtente(u);
+
+    final Email email_cancellazione = Email(
+      body: "L'attività prevista in data ${selectedDate.day}/${selectedDate.month}/${selectedDate.year}"
+          " dalle ore ${a.orainizio?.hour}: ${a.orainizio?.minute} alle ${a.orafine?.hour}: ${a.orafine?.minute} è stata assegnata ad un nuovo utente",
+      subject: 'cancellazione attività',
+      recipients: [u.email!],
+      isHTML: false,
+    );
+    FlutterEmailSender.send(email_cancellazione);
   }
 }
